Add files via upload
This is a JAVA based Housing lottery system developed by me and my team for Semester 3.
PS: We took this project not only as a project, but to solve the real world issue related to the housing schemes. This ideal solution along with some modifications can help out to eradicate or minimize the problems in the current existing system. I have included two PDFs in it so that a person accessing the file might get to know how its working.